workshop/code/requirements.txt
# core agent framework
strands-agents>=0.1.0
bedrock-agentcore>=0.1.0
strands-agents-tools>=0.1.0

# web automation
playwright>=1.40.0

# AWS integrations
langchain-aws>=0.1.0
boto3>=1.34.0
botocore>=1.34.0